Move cross-canister call gen to service module
