+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/service/rust/post_await_state_management.rs
