+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/service/rust/pre_await_state_management.rs
