+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/service/rust/promise_fulfillment.rs
