+++ src/compiler/typescript_to_rust/azle_generate/src/generators/ic_object/functions/call_raw128.rs
@@ -1,4 +1,4 @@
-use crate::generators::cross_canister_calls::{
+use crate::candid_type::service::rust::{
     post_await_state_management, pre_await_state_management, promise_fulfillment,
 };
