+++ src/compiler/typescript_to_rust/azle_generate/src/generators/mod.rs
@@ -1,2 +1,1 @@
-pub mod cross_canister_calls;
 pub mod ic_object;
